+++ src/metrepl/middleware/op_metrics.clj
@@ -6,20 +6,18 @@
   (:import
    [java.lang.management ManagementFactory]))
 
-(defonce first-op-metrified?* (atom false))
+(defonce first-op-received?* (atom false))
 
 (defn wrap-op-metrics
   "Wrap all ops metrifying each one, emmiting these events:
-   - `:event/first-op-requested` when receiving the first op.
+   - `:info/repl-ready` when receiving the first op metrifying project/repl info.
    - `:event/op-requested` when receiving any op.
    - `:event/op-completed` after completing an op."
   [handler]
   (fn [msg]
-    (when-not @first-op-metrified?*
-      (reset! first-op-metrified?* true)
-      (metrics/metrify :event/first-op-requested
-                       {:op (:op msg)
-                        :startup-time-ms (.getUptime (ManagementFactory/getRuntimeMXBean))}))
+    (when-not @first-op-received?*
+      (reset! first-op-received?* true)
+      (metrics/metrify-repl-ready (.getUptime (ManagementFactory/getRuntimeMXBean))))
     (handler (metrics/metrify-op-task msg))))
 
 (middleware/set-descriptor!